Fix module-size lint failure
